Adding a couple libraries to help with the flattening
